+++ sample/src/app/messaging/messaging.component.ts
@@ -1,11 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { AngularFireMessaging } from '@angular/fire/messaging';
 import { trace } from '@angular/fire/performance';
-import { Observable, fromEvent } from 'rxjs';
-import { tap, map, filter } from 'rxjs/operators';
-import * as firebase from 'firebase/app';
-import 'firebase/messaging';
+import { Observable } from 'rxjs';
+import { tap } from 'rxjs/operators';
 import { SwPush } from '@angular/service-worker';
+import { environment } from '../../environments/environment';
 
 @Component({
   selector: 'app-messaging',
@@ -25,14 +24,23 @@ export class MessagingComponent implements OnInit {
   message$: Observable<any>;
   showRequest = false;
 
-  constructor(public readonly messaging: AngularFireMessaging, swPush: SwPush) {
-    messaging.usePublicVapidKey('BIDPctnXHQDIjcOXxDS6qQcz-QTws7bL8v7UPgFnS1Ky5BZL3jS3-XXfxwRHmAUMOk7pXme7ttOBvVoIfX57PEo').then(() => {
+  constructor(public readonly messaging: AngularFireMessaging, readonly swpush: SwPush) {
+    swpush.messages.subscribe(it => console.log('swpush', it));
+    /*
+    TODO get this sorted back out with Firebase 8
+    messaging.usePublicVapidKey(environment.vapidKey).then(async () => {
+      if (navigator && navigator.serviceWorker) {
+        const registration = await navigator.serviceWorker.getRegistration();
+        if (registration) {
+          await messaging.useServiceWorker(registration);
+        }
+      }
       this.message$ = messaging.messages;
       this.token$ = messaging.tokenChanges.pipe(
         trace('token'),
         tap(token => this.showRequest = !token)
       );
-    });
+    });*/
   }
 
   ngOnInit(): void {