+++ sample/src/main-sw.js
@@ -1,0 +1,2 @@
+importScripts('/ngsw-worker.js');
+importScripts('/firebase-messaging-sw.js');