+++ src/auth-guard/auth-guard.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable, NgZone, Optional } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
 import { Observable, of, pipe, UnaryFunction } from 'rxjs';
 import { map, observeOn, shareReplay, switchMap, take, tap } from 'rxjs/operators';
-import { User } from 'firebase/app';
+import firebase from 'firebase/app';
 import {
   ɵAngularFireSchedulers,
   FirebaseOptions,
@@ -14,7 +14,7 @@ import {
 } from '@angular/fire';
 
 export type AuthPipeGenerator = (next: ActivatedRouteSnapshot, state: RouterStateSnapshot) => AuthPipe;
-export type AuthPipe = UnaryFunction<Observable<User|null>, Observable<boolean|any[]>>;
+export type AuthPipe = UnaryFunction<Observable<firebase.User|null>, Observable<boolean|string|any[]>>;
 
 export const loggedIn: AuthPipe = map(user => !!user);
 
@@ -23,7 +23,7 @@ export const loggedIn: AuthPipe = map(user => !!user);
 })
 export class AngularFireAuthGuard implements CanActivate {
 
-  authState: Observable<User|null>;
+  authState: Observable<firebase.User|null>;
 
   constructor(
     @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
@@ -38,14 +38,13 @@ export class AngularFireAuthGuard implements CanActivate {
     const auth = of(undefined).pipe(
       observeOn(new ɵAngularFireSchedulers(zone).outsideAngular),
       switchMap(() => zone.runOutsideAngular(() => import('firebase/auth'))),
-      tap((it: any) => it), // It seems I need to touch the import for it to do anything... race maybe?
       map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
       map(app => zone.runOutsideAngular(() => app.auth())),
       shareReplay({ bufferSize: 1, refCount: false }),
     );
 
     this.authState = auth.pipe(
-      switchMap(auth => new Observable<User|null>(auth.onAuthStateChanged.bind(auth))),
+      switchMap(auth => new Observable<firebase.User|null>(auth.onAuthStateChanged.bind(auth))),
       keepUnstableUntilFirst
     );
   }
@@ -55,24 +54,33 @@ export class AngularFireAuthGuard implements CanActivate {
     return this.authState.pipe(
       take(1),
       authPipeFactory(next, state),
-      map(can => typeof can === 'boolean' ? can : this.router.createUrlTree(can as any[]))
+      map(can => {
+        if (typeof can === 'boolean') {
+          return can;
+        } else if (Array.isArray(can)) {
+          return this.router.createUrlTree(can);
+        } else {
+          // TODO(EdricChan03): Add tests
+          return this.router.parseUrl(can);
+        }
+      })
     );
   }
 
 }
 
 export const canActivate = (pipe: AuthPipeGenerator) => ({
-    canActivate: [ AngularFireAuthGuard ], data: { authGuardPipe: pipe }
+  canActivate: [ AngularFireAuthGuard ], data: { authGuardPipe: pipe }
 });
 
 
 export const isNotAnonymous: AuthPipe = map(user => !!user && !user.isAnonymous);
-export const idTokenResult = switchMap((user: User|null) => user ? user.getIdTokenResult() : of(null));
+export const idTokenResult = switchMap((user: firebase.User|null) => user ? user.getIdTokenResult() : of(null));
 export const emailVerified: AuthPipe = map(user => !!user && user.emailVerified);
 export const customClaims = pipe(idTokenResult, map(idTokenResult => idTokenResult ? idTokenResult.claims : []));
 export const hasCustomClaim: (claim: string) => AuthPipe =
   (claim) => pipe(customClaims, map(claims =>  claims.hasOwnProperty(claim)));
-export const redirectUnauthorizedTo: (redirect: any[]) => AuthPipe =
+export const redirectUnauthorizedTo: (redirect: string|any[]) => AuthPipe =
   (redirect) => pipe(loggedIn, map(loggedIn => loggedIn || redirect));
-export const redirectLoggedInTo: (redirect: any[]) => AuthPipe =
+export const redirectLoggedInTo: (redirect: string|any[]) => AuthPipe =
   (redirect) => pipe(loggedIn, map(loggedIn => loggedIn && redirect || true));