+++ src/functions/functions.ts
@@ -9,15 +9,18 @@ import {
   ɵAngularFireSchedulers,
   ɵfirebaseAppFactory,
   ɵlazySDKProxy,
-  ɵPromiseProxy
+  ɵPromiseProxy,
+  ɵapplyMixins
 } from '@angular/fire';
-import { functions } from 'firebase/app';
+import firebase from 'firebase/app';
+import { proxyPolyfillCompat } from './base';
+import { HttpsCallableOptions } from '@firebase/functions-types';
 
 export const ORIGIN = new InjectionToken<string>('angularfire2.functions.origin');
 export const REGION = new InjectionToken<string>('angularfire2.functions.region');
 
 // override httpsCallable for compatibility with 5.x
-export interface AngularFireFunctions extends Omit<ɵPromiseProxy<functions.Functions>, 'httpsCallable'> {
+export interface AngularFireFunctions extends Omit<ɵPromiseProxy<firebase.functions.Functions>, 'httpsCallable'> {
 }
 
 @Injectable({
@@ -39,7 +42,6 @@ export class AngularFireFunctions {
     const functions = of(undefined).pipe(
       observeOn(schedulers.outsideAngular),
       switchMap(() => import('firebase/functions')),
-      tap((it: any) => it),
       map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
       map(app => app.functions(region || undefined)),
       tap(functions => {
@@ -50,10 +52,10 @@ export class AngularFireFunctions {
       shareReplay({ bufferSize: 1, refCount: false })
     );
 
-    this.httpsCallable = <T = any, R = any>(name: string) =>
+    this.httpsCallable = <T = any, R = any>(name: string, options?: HttpsCallableOptions) =>
       (data: T) => from(functions).pipe(
         observeOn(schedulers.insideAngular),
-        switchMap(functions => functions.httpsCallable(name)(data)),
+        switchMap(functions => functions.httpsCallable(name, options)(data)),
         map(r => r.data as R)
       );
 
@@ -62,3 +64,5 @@ export class AngularFireFunctions {
   }
 
 }
+
+ɵapplyMixins(AngularFireFunctions, [proxyPolyfillCompat]);