+++ src/performance/performance.module.ts
@@ -12,6 +12,6 @@ export class AngularFirePerformanceModule {
   ) {
     // call anything here to get perf loading
     // tslint:disable-next-line:no-unused-expression
-    perf.dataCollectionEnabled;
+    perf.dataCollectionEnabled.then(() => {});
   }
 }