+++ src/schematics/deploy/actions.ts
@@ -5,8 +5,6 @@ import { copySync, removeSync } from 'fs-extra';
 import { dirname, join } from 'path';
 import { execSync } from 'child_process';
 import { defaultFunction, defaultPackage, NODE_VERSION } from './functions-templates';
-import { experimental } from '@angular-devkit/core';
-import { SchematicsException } from '@angular-devkit/schematics';
 import { satisfies } from 'semver';
 import * as open from 'open';
 
@@ -21,7 +19,8 @@ const deployToHosting = (
   firebaseTools: FirebaseTools,
   context: BuilderContext,
   workspaceRoot: string,
-  preview: boolean
+  preview: boolean,
+  firebaseToken?: string,
 ) => {
 
   if (preview) {
@@ -42,7 +41,8 @@ const deployToHosting = (
         return firebaseTools.deploy({
           // tslint:disable-next-line:no-non-null-assertion
           only: 'hosting:' + context.target!.project,
-          cwd: workspaceRoot
+          cwd: workspaceRoot,
+          token: firebaseToken,
         });
       } else {
         return Promise.resolve();
@@ -54,7 +54,8 @@ const deployToHosting = (
     return firebaseTools.deploy({
       // tslint:disable-next-line:no-non-null-assertion
       only: 'hosting:' + context.target!.project,
-      cwd: workspaceRoot
+      cwd: workspaceRoot,
+      token: firebaseToken,
     });
 
   }
@@ -117,9 +118,11 @@ export const deployToFunction = async (
   firebaseTools: FirebaseTools,
   context: BuilderContext,
   workspaceRoot: string,
-  project: experimental.workspace.WorkspaceTool,
+  staticBuildTarget: BuildTarget,
+  serverBuildTarget: BuildTarget,
   preview: boolean,
-  fsHost: FSHost = defaultFsHost
+  firebaseToken?: string,
+  fsHost: FSHost = defaultFsHost,
 ) => {
   if (!satisfies(process.versions.node, getVersionRange(NODE_VERSION))) {
     context.logger.warn(
@@ -127,30 +130,22 @@ export const deployToFunction = async (
     );
   }
 
-  if (
-    !project ||
-    !project.build ||
-    !project.build.options ||
-    !project.build.options.outputPath
-  ) {
-    throw new SchematicsException(
-      `Cannot read the output path (architect.build.options.outputPath) of the Angular project in angular.json`
+  const staticBuildOptions = await context.getTargetOptions(targetFromTargetString(staticBuildTarget.name));
+  if (!staticBuildOptions.outputPath || typeof staticBuildOptions.outputPath !== 'string') {
+    throw new Error(
+      `Cannot read the output path option of the Angular project '${staticBuildTarget.name}' in angular.json`
     );
   }
 
-  if (
-    !project ||
-    !project.server ||
-    !project.server.options ||
-    !project.server.options.outputPath
-  ) {
-    throw new SchematicsException(
-      `Cannot read the output path (architect.server.options.outputPath) of the Angular project in angular.json`
+  const serverBuildOptions = await context.getTargetOptions(targetFromTargetString(serverBuildTarget.name));
+  if (!serverBuildOptions.outputPath || typeof serverBuildOptions.outputPath !== 'string') {
+    throw new Error(
+      `Cannot read the output path option of the Angular project '${serverBuildTarget.name}' in angular.json`
     );
   }
 
-  const staticOut = project.build.options.outputPath;
-  const serverOut = project.server.options.outputPath;
+  const staticOut = staticBuildOptions.outputPath;
+  const serverOut = serverBuildOptions.outputPath;
   const newClientPath = join(dirname(staticOut), staticOut);
   const newServerPath = join(dirname(serverOut), serverOut);
 
@@ -204,7 +199,8 @@ export const deployToFunction = async (
     return firebaseTools.deploy({
       // tslint:disable-next-line:no-non-null-assertion
       only: `hosting:${context.target!.project},functions:ssr`,
-      cwd: workspaceRoot
+      cwd: workspaceRoot,
+      token: firebaseToken,
     });
   }
 };
@@ -212,13 +208,15 @@ export const deployToFunction = async (
 export default async function deploy(
   firebaseTools: FirebaseTools,
   context: BuilderContext,
-  projectTargets: experimental.workspace.WorkspaceTool,
-  buildTargets: BuildTarget[],
+  staticBuildTarget: BuildTarget,
+  serverBuildTarget: BuildTarget | undefined,
   firebaseProject: string,
-  ssr: boolean,
-  preview: boolean
+  preview: boolean,
+  firebaseToken?: string,
 ) {
-  await firebaseTools.login();
+  if (!firebaseToken) {
+    await firebaseTools.login();
+  }
 
   if (!context.target) {
     throw new Error('Cannot execute the build target');
@@ -226,10 +224,16 @@ export default async function deploy(
 
   context.logger.info(`📦 Building "${context.target.project}"`);
 
-  for (const target of buildTargets) {
+  const run = await context.scheduleTarget(
+    targetFromTargetString(staticBuildTarget.name),
+    staticBuildTarget.options
+  );
+  await run.result;
+
+  if (serverBuildTarget) {
     const run = await context.scheduleTarget(
-      targetFromTargetString(target.name),
-      target.options
+      targetFromTargetString(serverBuildTarget.name),
+      serverBuildTarget.options
     );
     await run.result;
   }
@@ -255,20 +259,23 @@ export default async function deploy(
       })
     );
 
-    if (ssr) {
+    if (serverBuildTarget) {
       await deployToFunction(
         firebaseTools,
         context,
         context.workspaceRoot,
-        projectTargets,
-        preview
+        staticBuildTarget,
+        serverBuildTarget,
+        preview,
+        firebaseToken,
       );
     } else {
       await deployToHosting(
         firebaseTools,
         context,
         context.workspaceRoot,
-        preview
+        preview,
+        firebaseToken,
       );
     }