+++ src/schematics/interfaces.ts
@@ -9,6 +9,7 @@ export interface Project {
 export interface FirebaseDeployConfig {
   cwd: string;
   only?: string;
+  token?: string;
 }
 
 export interface FirebaseTools {
@@ -73,4 +74,14 @@ export interface FSHost {
   moveSync(src: string, dest: string): void;
   writeFileSync(src: string, data: string): void;
   renameSync(src: string, dest: string): void;
+}
+
+export interface WorkspaceProject {
+  projectType?: string;
+  architect?: Record<string, { builder: string; options?: Record<string, any> }>;
+}
+
+export interface Workspace {
+  defaultProject?: string;
+  projects: Record<string, WorkspaceProject>;
 }